Removing All Gem Installs and Unneeded Files
This commit removes the unused Jekyll functionality of this gem by
deleting all the gem installs except the dependency defined in the
gemspec. It also removes the configuration and documentation files
that were only used for creating standalone sites. atlas_assets is
now stripped down to only be used for packaging the assets files in
the atlas frontend.